some changes

+++ simulated-agent-model/gifs.py
@@ -22,13 +22,14 @@ def makeGif(sigmaHat, rationalizationFactor, numVoters, iterations):
     print("compiling voter curve gif...")
     name = '../figs/simulation-figs/gifs/votercurve-{}-{}.gif'.format(sigmaHat,rationalizationFactor)
     with io.get_writer(name, mode='I') as writer:
-        for filename in filelist:
+        for idx, filename in enumerate(filelist):
             image = io.imread(filename)
             writer.append_data(image)
-            os.remove(filename)
+            if idx%10 != 0:
+                os.remove(filename)
 
-    # Remove files
-    os.rmdir('../figs/simulation-figs/gif-{}-{}-{}'.format(sigmaHat, rationalizationFactor,numVoters))
+    # # Remove files
+    # os.rmdir('../figs/simulation-figs/gif-{}-{}-{}'.format(sigmaHat, rationalizationFactor,numVoters))
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser()
